Force screenshot sound when camera view is active
Force the camera shutter sound when taking a screenshot while the camera
is open if the camera sound is forced.

The camera sound is forced in certain regions but the camera shutter
sound when taking a screenshot respects system mute. This change forces
the camera shutter sound when taking a screenshot while the camera is
open if the camera sound is forced. This is to remove the possibility to
work around the forced camera sound by taking a screenshot of the camera
viewfinder.

Flag: com.android.systemui.force_screenshot_sound_if_must_play_shutter_sound
Bug: 329483874
Test: atest ScreenshotSoundControllerTest
Test: 1) $ adb shell setprop audio.camerasound.force true
      2) Open the camera and take a screenshot.
         Verify that the forced camera sound is played when the camera
         is open, regardless of system volume setting.
Change-Id: Ide6cc211a3b1eee9893604bbe16b5dd8cba6d7c0