USOAC: Fix system animation disable check
Settings sets the animator_duration_scale value to a float value of 0.0.
Getting its value as a string returns '0.0' not '0'.
So the check is always false here.

Because of this, UnlockedScreenOffAnimationController tries to
start the AOD screen animation even if the system animation disabled.
It makes the AOD screen goes black or just status bar visible due to the
animation flow messed up.
It's even worse and clearly noticeable when screen lock is set to none.

Test:
1. Set screen lock to 'None'
2. Turn off system animations in developer settings.
3. Enable AOD.
4. Turn off screen via power button.
5. AOD screen is visible correctly

Change-Id: Icc3753c8cb462c1233b418095e4129eb992c1f9c