Storage: Add speedClass property to DiskInfo for an SDCard (2/2)
Currently, there is no property persent for SDCard speed class of an
SDCard.

Having it helps the OEM framework in
a) notifying the user of performance impact if a lower class SD Card
   is inserted.
b) Capture metrics related to Class of SD Card.

To get the speed class the third byte of the ssr is read and is
right shifted by 24. The decimal value corresponds to the following -

0: Class 0
1: Class 2
2: Class 4
3: Class 6
4: Class 10

and is added to the DiskInfo via onDiskMetadataChanged method

Test: manual
1. Install an SDCard and note down the speed class on the card.
2. Ran adb shell dumpsys mount and checked the speedClass property in
   the DiskInfo section

Change-Id: I28e19fd87753bd9a0fca495e6aae84cd5ca2c255