Prevent error message on svc power reboot
With the changes to sys.powerctl this variable is no longer set on a
power event - the power action is simply carried out. (This is to avoid
possible DoS of these power functions.) However, this means that this
property can no longer read to see if a power event is happening. Since
there is no longer a way of telling if a remote exception is real or
not, best not to print out an error.

Note that this bug occurs on some devices, not all. I was able to
reproduce it reliably on a Pixel 7a, but not on a Pixel 7 or Cuttlefish.

Bug: 318323013
Test: adb shell svc power reboot

Change-Id: Icbc1ec1b7882140d4d742b056ea2adaafae67ee9