Launch target activity on power key double tap
With this change, if the double press on power key camera gesture is
disabled the power key double press gesture will not be intercepted in
GestureLauncherService but instead handled by the single key gesture
detector in PhoneWindowManager.

If config_doublePressOnPowerBehavior is set to 3, the default activity
to launch will be config_defaultDoublePressOnPowerTargetActivity. This
change also adds a new secure setting so the target activity can be
changed.

Also implement the dismissKeyguardToLaunch method in
KeyguardViewMediator.

Bug: 327818922
Test: atest GestureLauncherServiceTest
Change-Id: I4dac90644054abd45460ee7a36d001b928591a36