SystemUI: token_alignment: Fix vulnerabilities found by npm
We can find vulnerabilities with "npm audit".
And then running "npm audit fix" to generate this CL.

The index.ts depends on a non-existed directory in my environment, so I
can't test it locally, and leave the validation to TH.

Another interesting fact: GitHub's dependabot alos find these
vulnerabilities and alert us if we place the frameworks/base source code
on GitHub.

Test: TH
Bug: 328710557

Change-Id: Ice61c3f058d18cca89cbf48db00fb42e6548389a
Signed-off-by: utzcoz <utzcoz@gmail.com>