Revert "Move frameworks/base/Android.mk to build/tasks"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9461f2f6-62f3-4bdd-8206-a4d57e5049c7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iafd9999070a4ecacc1f6b38ce48109a6e6bedd88