Moves KeyStore key management for StorageAreas out of StorageManager
Responding to API review: making the StorageManager API for StorageAreas
more general, so `openStorageArea` takes a byte[] secret instead of using
the KeyStore to generate an UnlockedDeviceRequired key that was then used
to derive the secret.

Also renames OpenStorageArea class to StorageArea, and adds some death management so that StorageAreas are closed when garbage collected or on app crash.

Change-Id: I11e96884364389dd155d868b494d97dff5c66b50