Revert "Workaround build breakage from full LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/90009d39-c98a-4883-baa8-f2bed5aa8373). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I59bf5073bae67abbb9e2d93090f163ad41dff7e9