Revert "Workaround build breakage from full LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3fed3a3c-a1b4-4625-86a0-bb0a98b5075c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib14c7d749b6df0367e23928bc3db330512af8d79