Response stats counts are not updated sometimes
This problem occurs only when the timestamp of broadcast and the
notification event is exactly the same.
Though ideally they need to occur one after the other,
it is quite possible that they can occur immediately with the same
timestamp. Hence, we see flaky behaviour.
Fix is to process the events even if broadcast event's
timestamp is equal to the notification event.

Bug: 330844354

Test: atest tests/tests/app.usage/src/android/app/usage/cts/BroadcastResponseStatsTest.java
Change-Id: I69398c1d15ed7a4ae39c31a2b56f19ef0b858614