Revert "Update LocaleUtils.filterByLanguage due to addLikelySubt..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/82cec203-9df9-4d33-a046-d9fa6a2fd6e1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iae3d58dc603449da0fd4a0227a53cfc218f965a9