PM: Make ApplicationInfo#IS_DATA_ONLY functional (1/1)
Currently, ApplicationInfo#IS_DATA_ONLY is not being used by any of the
APIs from PackageManager. PackageManager allows packages to uninstall
while preserving the data of the package using DELETE_KEEP_DATA flag.
Today there is no way to determine whether the package was deleted
using the flag or not.

This CL (re) introduces the use of the flag for the packages which were
uninstalled using the flag DELETE_KEEP_DATA. This CL adds new method in
PackageSetting and PackageStateInternal to keep track of the state.
This is set in RemovePackageHelper if the uninstall flag is set.
This CL also prevents removal of the package during package scan if the
APK file is absent.
The usage of the flag was removed with the commit:
d6ee0ba57906f83f45ca2a2cd835d7ff6e7a88df

Test: manual
1. Install a APK in the device.
2. Execute adb shell cmd package uninstall -k <package_name>
3. Execute adb shell dumpsys package <package_name> and check the
   IS_DATA_ONLY flag is set for the package.
4. Reboot the device, repeat step 3.

Also ran, atest PackageManagerServiceTests and checked there are no
new failures are there.

Change-Id: Ibeda3a89ef7c19ce49d0da9891369f762bd45458