Revert "Update LocaleUtils.filterByLanguage due to addLikelySubt..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/89bf666c-907b-43dc-b47c-12187659630d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9a69797ddac3c5c9ad1bc7ec5ba3258af0dac21f