Revert "Update LocaleUtils.filterByLanguage due to addLikelySubt..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/89bf666c-907b-43dc-b47c-12187659630d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If3de4e2f4c5bb07f7ca1553122161e9f745422ed