Revert "Updating ownership of color resources"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1c6d1b39-1dd4-4a76-99be-94530fb0aef1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5c30fa4afb617d6cd591d4551221303c251c486c