Revert "Updating ownership of color resources"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ad2fa51f-0904-4f25-a87a-a75b6e1b7a4c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia4d53b72a537b4d135577f89d9a32fc0933878ce