Fix potential crash issue in NonRequiredPackageDeleteObserver
mLatch.notifyAll() will not help wake the thread which is blocked by
mLatch.await, instead, it will cause system server crash.

So, mark state as failed if any package is failed to be deleted, and let
mLatch.countDown work until it comes to 0 and wake the thread blocked by
await.

Change-Id: Ida1c3c128ddef376aaaacda1358717d0d4e9ca8b