If the dialog is already above the IME window, there is no need to set a relative hierarchy.
If you want the current system dialog to be above the IME window, it is only appropriate to set the relative level when the current dialog level is lower than the IME window. If the current dialog is already above the IME window, setting the relative hierarchy is likely to lower the current window's hierarchy, causing other windows that should have been below it to appear above the current window, causing a hierarchy disorder problem.

Bug: 336409037
Test:

Change-Id: I7848a762642566bb9cd27671eed35dc29afdebe0
Signed-off-by: wanghuadong <wanghuadong@xiaomi.corp-partner.google.com>