Inform Listeners when initializeDefaultsForSystemsApps() change idleness buckets
Note: This is an upstream from Meta.

Problem:
initializeDefaultsForSystemApps() is not informing listeners of change in uid idleness.
This results in stale idleness values in components such as the network policy manager, which
has listeners to uid changes.

Solution:
call maybeInformListeners() to make sure listeners are informed of the changes

Test: Manual test on my own Meta device.

 Tested on a race condition where network policy manager queried for the idleness of a uid before
 initializeDefaultsForSystemApps() is called.

 Before the fix, network policy manager got STANDBY_BUCKET_NEVER as idleness
 and never got informed of change in uid when initializeDefaultsForSystemApps()
 sets uid idleness bucket to STANDBY_BUCKET_ACTIVE

 After the fix, network policy manager's listener was able to be informed of this change in uid.

Change-Id: Ic6b665eea4a07d133a85789bce1a1af514e7ee96