Revert "Add metaData null check in isRequiredAppDeclaredInMetadata"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cb21f2d9-5a8f-4d75-b72f-5e6f3f5cc8fb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I526e8ff0c92a0619089316b407a4267e549c472a