Encode IkeSessionParams using getIkeOptions
This patch integrates IkeSessionParams#getIkeOptions into
the process of making IkeSessionParams a PersistableBundle. With
this patch, the encoding process no longer depends on a static
IKE_OPTIONS list. Thus if a new IKE Option is added in the IKE
module, it will be taken care of without needing to update the
IKE_OPTIONS list

Bug: 185941731
Test: atest FrameworksVcnTests && atest CtsVcnTestCases
Flag: NONE mainline API flag not accessible; using try catch
      block to gate the behavior
Change-Id: I2c1a9ca1731362b1ab666e5b804911d8b0ca45e7