Add Option to Magnify Navigation Bar and IME [1/2]
Users who use screen magnifier now have the option to magnify
their navigation bar and IME (e.g. keyboard). We implement this through
a new Secure Setting monitored through the AccessibilityManagerService,
which can be toggled by a user in the Settings.

Test: Manual
Navigate to Settings > Accessibility > Magnification
and turn on Magnify Navigation Bar and Keyboards .
Test that the navigation bar and keyboard can be magnified
depending on the setting. Turn the setting off, and the nav bar
and keyboard should not be magnifiable.

Flag: ACONFIG android.view.accessibility.magnify_nav_and_ime DEVELOPMENT

Change-Id: I5496bcd9b5c3cf94620eedc7c86d12f31e86889a