Revert "Simplify MediaSessionRecord callback invocations"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7371ab60-fdd3-4de2-99e0-c990b4e0a3e1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I120e9b79fdfda393cdd691b8b2a092fb38e20524