Revert "Fix typo in aapt2_results genrule."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c3903057-4871-44d6-8a75-8b40ac2ef85d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8ad1effbda7232cb2127a51882060b8ddf0c7db1