Revert "Fix typo in aapt2_results genrule."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4d71929-494a-483a-b6a1-bb1058352852). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie7dec443ff5e3723cf9cc0b72b7903b25d5eee36