Stop capturing a reference to LockscreenCredential
This CL changes the `setLockCredentialInternal` in the
LockSettingsService so that it no longer stores a reference to the
LockscreenCredential object passed in, but instead stores
a duplicate.

This change means that the caller can safely zeroize the
LockscreenCredential it passes in, which is particularly important
since it allows users to use try-with-resources with their local
LockscreenCredential.

Bug: 338122208
Test: atest CtsDevicePolicyTestCases -- --abi arm64-v8a
Change-Id: I8b4f97524f71bd66753e09260c6ae0db20322f49