Stop capturing a reference to LockscreenCredential
This CL changes the `setLockCredential` in the
LockSettingsService so that it passes a duplicate credential to
`notifyPasswordChanged`, since `notifyPasswordChanged` captures the
credential in the scope of its callback. Then, it also
zeroizes the credential in `notifyPasswordChanged` after use.

This change means that the caller can safely zeroize the
LockscreenCredential it passes in, which is particularly important
since it allows users to use try-with-resources with their local
LockscreenCredential.

This CL also adds a finalizer to the `KeySyncTask`, to zeroize its
`byte[] mCredential` field storing the LSKF.

Bug: 338122208
Test: atest CtsDevicePolicyTestCases -- --abi arm64-v8a
Change-Id: I8b4f97524f71bd66753e09260c6ae0db20322f49