Use kN32_SkColorType when creating the buffer for SkiaHostPipeline
kN32_SkColorType is determined at compile-time using a macro that checks
the current architectiure. On Mac, this value is different than Linux or
Windows.  As a result, when HW rendering occurs on a Mac, the blue and
red pixels are swapped, which implies one is using
kRGBA_8888_SkColorType and the other is using kBGRA_8888_SkColorType.

When an Android Bitmap object is initialized using
Bitmap.Config.ARGB_8888, it uses kN32_SkColorType

Update SkiaHostPipeline to use kN32_SkColorType so the value is
consistent with what Bitmap uses.

Test: HardwareAcceleratedActivityRenderTest on Mac
Test: ShadowNativeHardwareRendererTest on Mac

Bug: 338399395
Change-Id: Ic037fea16abcbf950d69daafadbf91fb9c6653a6