Revert "Fix typo in aapt2_results genrule."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2e1b7b8c-1e91-4c58-b640-7d2188e4d162). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I547bb2d0d8fc0a6f521b37d615c758d01fb832eb