DeviceIdleController: synchronize dumpTempWhitelistSchedule
dumpTempWhitelistSchedule is not synchronized so it can potentially access mTempWhitelistAppIdEndTimes while an existing operation to the list is in progress.

There are two issues with this:
1) mTempWhitelistAppIdEndTimes is a SparseArray which is not a thread safe class.
2) An operation to add the app to the whitelist may not have completed when a request
to dump the list is received.

The specific use-case is the CTS test case android.alarmmanager.cts.ExactAlarmsTest#alarmClockGrantsWhitelist where the whitelist is checked right after an alarm fires.On most AOSP devices,there is no problem
because the whitelist addition completes before the app dumps thewhitelist. But let's say if there is additional activity in DeviceIdleController which might cause whitelist addition to take slightly longer then the test might fail.

We already synchronize a large block in DeviceIdleController.dump, so there should be minimal risk in also synchronizing DeviceIdleController.dumpTempWhitelistSchedule.Both of these methods are used for debugging.

Test: atest --iterations 5 android.alarmmanager.cts.ExactAlarmsTest#alarmClockGrantsWhitelist
Change-Id: I1f032cb3e983ecf75e5607d9510a6f8ffee13607
Signed-off-by: Abhishek Gadewar <abhishekgadewar@meta.com>