Revert "Support properties in path when merging overlay configs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/998d6004-ea76-4d76-872b-cf0e8a64edd7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 341139531
Change-Id: Ib0aa1dd6dbe2387e02a53731a0fc9958f2891b0f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com