Update to not set motion alarm for TV device
DeviceIdleController's motion alarm wake up 10-11 minutes after AC on.
The alarm ends without any processing, but it is an unnecessary alarm for the TV,
and in order to comply with ECO regulations(Power Consumption for EU),
we changed it so that the motion alarm does not start when the motion
sensor is not used.

Test: Manual
Change-Id: Ic53c3d4f9323f1b13fb0344a979d3967f8c6133c