Use full APIs for api_lint.new_since of *api-stubs-docs-non-updatable
Previously, the `*api-stubs-docs-non-updatable` modules used either the
full previously released public API or for those API surfaces that
extend another API surface the delta to that surface. That meant that
lint would not be able to properly ignore issues that already existed
in previously released APIs.

This change switches them to use the `*.api.combined.*.latest` modules
which contains the full API for each surface, including the API
provided by any API surface it extends. That ensures that lint will
not report issues with previously released APIs.

Bug: 341078916
Test: ./gradlew
Change-Id: I8a6d1263086466fda111266fe8bfdfedf0f8c152