Revert "Publish TracingServiceProxy after storage has been unlocked"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/96152cd5-44a1-4042-a434-d777c42d460b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6903aa1b0e3608d4c6babf20f5bf41158e200243
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com