BinderProxy: tryDrop
The benefit: instead of re-using callbacks or binder
objects, this allows you to close them. Closing a
binder releases native resources (which will also
remove the binder from the native binder proxy count,
and avoid triggering the too many binder proxies
allocated area). So, remote resources will be
freed, and you can wait on the garbage collector
to delete resources in-process.

Test: atest android.os.BinderProxyTest
Bug: many

More technical details on why implementation here:

BpBinder is OBJECT_LIFETIME_WEAK, and there are many
lockless accesses of it, so even if you figure out
when to delete it, there's no way to delete it and
safely access it. Modifying RefBase or BpBinder may
be possible, but it would be exceedingly complex.

Doing this in:
- RefBase: without adding a new object, maybe you
  could do it with forceDecStrong, but it would
  break all the other logic assuming you can promote
  OBJECT_LIFETIME_WEAK, so we could add an additional
  flag 'OLW but deleted anyway' and then the Java code
  is closing and affecting native code that may also
  be using the binder and assuming it's OLW externally
  to RefBase, and it's dangerous since we don't have
  visibility on that.
- BpBinder: we could add close to BpBinder itself,
  but besides it being a not-best-practice-design,
  and some weird interactions with the destructor,
  onLastStrongRef, onIncStrongAttempted, it may be
  okay, and can try it as an alternative.

The way this CL works around this is to create a new
allocation per BinderProxy object so that we can
safely own and delete the underlying BpBinder. This is
pretty expensive, may need to be cached. However, it
is the cleanest/most separation of concerns approach.

However, you can delete proxies like this.

Change-Id: I18edba8f296044f1b0d0bdd29d570f06c5bfd99e