BinderProxy close
Test: WIP !! (no time)
Bug: many

BpBinder is OBJECT_LIFETIME_WEAK, and there are many
lockless accesses of it, so even if you figure out
when to delete it, there's no way to delete it and
safely access it. Modifying RefBase or BpBinder may
be possible, but it would be exceedingly complex.

Doing this in:
- RefBase: without adding a new object, maybe you
  could do it with forceDecStrong, but it would
  break all the other logic assuming you can promote
  OBJECT_LIFETIME_WEAK, so we could add an additional
  flag 'OLW but deleted anyway' and then the Java code
  is closing and affecting native code that may also
  be using the binder and assuming it's OLW externally
  to RefBase, and it's dangerous since we don't have
  visibility on that.
- BpBinder: we could add close to BpBinder itself,
  but besides it being a not-best-practice-design,
  and some weird interactions with the destructor,
  onLastStrongRef, onIncStrongAttempted, it may be
  okay, and can try it as an alternative.

So (this CL) one ne way around this is to create a new
allocation per BinderProxy object. This is pretty
expensive, may need to be cached.

However, you can delete proxies like this.

Change-Id: I18edba8f296044f1b0d0bdd29d570f06c5bfd99e