Allow special UIDs (SHELL, ROOT, etc) to observer content Uris
Summary: The permission checking logic for registering a content observer for a uri assumes that all processes are started by the activity manager, which is not the case. An example is a native unit test (gtest) that reaches out to register a content observer. The check if a process is in the known by the activity manager is purely for logging pruposes and not a security check which are done via permissions. We are now allowing special UIDs, as also handled by the AppOpsService to register observer. Technically, we should be able to support all UIDs given this is for logging but for now conservatively support the platform defined special UIDs.

Test: build successful

Change-Id: I31b602af8ba5fbe123231204ba26952171e420e0
Signed-off-by: Abhishek Gadewar <abhishekgadewar@meta.com>