Revert "create setChargingPolicy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/182ac2db-e7f8-412f-aea3-47f7957dc757). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I24724c0a4436a7a24432be97066e8bdb22d8fe32
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com