Revert "Revert "Revert "Revert "Revert "Workaround build breakag..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6a6575b2-e582-4694-b3a3-cbc66a180df2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5afc879ac3e963b04cc4cc4cdb380feee30b206d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com