Revert "Added trace event to broadcastIntentWithFeature."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c589e25a-79e6-43cb-be9e-f654b9a7fb0c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 347743665
Change-Id: I75d68a442a29cfea1fefcf3a6c499a2f94e44dbc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com