Update OWNERS for art-profile* files
The current OWNERS file is outdated and does not include the latest members of the ART team. This CL updates the OWNERS file to include the latest members of the ART team and to use the OWNERS file in the platform/art directory as the source of truth for who should be an owner of these files.

Change-Id: Ib436179d49314dc8d43b04c2872737278741fd31