Enable use of app allocator defaults in forked.
If an app forks a process to do any work, it will revert back to the
allocator defauls which are more conservative and use a slightly slower
path. Set an environment variable that will cause a forked process to
use the same allocator settings as the app.

If the app overwrites all environment variables, this scheme will not work.

Test: Ran an app and verified a forked process uses the same allocator defaults.
Change-Id: I892940c27d407570557a006b0ac2afe93e166f11