Revert "SharedMemory: mark fdsan ownership"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3953fc9a-4f90-4d4c-93e4-ad965e4239d7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icc895a4b05ae0f406641bed274dc1cead74dce72
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com