Positioning settings modification log enhancements (1)
Log enhancement for third-party applications using setLocationEnabledForUser() functions in LocationManagerService.
This modification can easily identify which application has performed this operation.

We tested this function on the Xiaomi HyperOS (Android V), and the result log is as follows:

06-24 10:45:11.457: location [u0] disabled because com.android.systemui (setLocationEnabledForUser)
06-24 10:45:11.465: location [u0] disabled (onLocationModeChanged)
06-24 10:45:21.394: location [u0] disabled because com.android.settings
06-24 10:45:21.400: location [u0] disabled

To identify which application called setLocationEnabledForUser() function at what time and what operations were performed

Change-Id: Icb37edb0da315171ecbcdf04ee007a7bdaee1837