Fix the issue on all widgets without the directBootAware flag being lost
It's a common practice for Original Equipment Manufacturers (OEMs) to
start the launcher in advance (before the user unlocks), and the
stopListening method may also be called in advance.

If this call is made before the user unlocks, then according to the
existing processing logic of the framework (it returns directly if it's
already called), all widgets without the directBootAware flag will be
lost (can't be queried), and they can only be restored by restarting the
phone.

Test:

If ensureGroupStateLocked is called before the user unlocks, and the
second parameter is passed as false, the widget will be lost (on the
desktop or in the menu).

Change-Id: Ibf780a81fcbe8a0677aa8997892b3a242472d37f
Signed-off-by: Jia Jia <jia.jia@zte.com.cn>