Revert "Revert "Revert "Revert "Revert "Workaround build breakag..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/49d7c23a-8ad5-4e12-bb4b-acce9f54f2b1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 347840973
Change-Id: If6fcb42d4d9bcf4cb6f1f8af4c823aad490dd5f7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com