Revert "Revert "Revert "Revert "Revert "Workaround build breakag..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9a525f0d-9ebd-4753-83c5-c2b7b67c0853). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 347836801
Change-Id: Ifa13756fe54dd175bdafd8106aa0807506e280a6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com