Add better frameinfo data conversion logic to android_graphics_HardwareRenderer.cpp
Previously, in Android R and below, the 9-length frameinfo data was
copied directly into the 12-length frameinfo data from Android S.

However, the fields don't line up, as some fields were added at the
beginning of the data array in Android S.

Add some logic to convert from the 9-length to the 12-length. While the
conversion is missing some data, it appears to be fully functional.

Test: HW rendering tests in Google3
Bug: 344979295
Change-Id: I1631d08d9d9354d198d57a67dca39ea590b8055f