Revert "Create exported mode flag library for package android.os""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c73eab49-1978-48ab-969d-6d7873a86498). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 352485207
Change-Id: I18545a5d5f26803195d8818fe539b37618902a74
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com