Revert "Create exported mode flag library for package android.os""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/877fa06c-0242-47e7-9f2b-a25fdd4ca53e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 352485762
Change-Id: Ib5d30210aa9de046a8e999738ee9d5c49b47d0d6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com