AOSP BUG fix to utilize KEYCODE_APP_X as GlobalKey
KEYCODE_APP_X is expected as global key from Android to launch the
specific application for TV by following the rule of gloabalmanager.
Currently these keys are not delivered according to the global-key
implementation by OEMs. Making a change to ensure KEYCODE_APP_X is
delivered to OEM applications as expected by making the use of GlobalKeyManager.

Test: manual
Created local build to ensure functionality is working
Ran the test that ensures OEM apps are receiving the global-keys.

Change-Id: I073e76ece3ff631afef5e0614ef7bf34d3442f15
Signed-off-by: ONKAR KULKARNI <onkkul@amazon.com>