Revert "animateItemPlacement() is deprecated and needs to be upd..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/75a78785-0f51-4a25-9cf1-cc6b1e9ea363). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 353771362
Change-Id: I34445f5027cd0132bc6c9c3e39f1a078aea463fb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com