Revert "Support getting RPC service in java ServiceManager"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/13b914d2-d226-4a8c-93dc-b6c1b053a600). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I06824af14ae7d530a10d339b86411639ad76eaea
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com