Revert "Support getting RPC service in java ServiceManager"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2cbac161-b81c-495d-9fab-0215156ae502). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iad825af2cacbc1c6b0ff131208f4c0180b439e11
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com