Revert "Support getting RPC service in java ServiceManager"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d739370f-1264-4fb6-bffd-cafedf3d2fa9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia18e2c7df46ea0b2f02ddbb1c5075e43469a8cf7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com