Fix non-ASCII text garbled display issue
When the system language is Chinese, after attach a ringtone. The SoundPickerActivity's title garbled. It is caused by Html.escapeHtml does not support non-ASCII characters. Fix this by using Html.fromHtml instead.

Bug: 355350222
Test: Manual
Install the modified apk to see when the system language is Chinese,whether the text display normally or not.

Change-Id: I374bd320b106fc6f9785464dbfccef3e250403a1