Mark ProcessRecord as persistent for child processes of persistent apps
Service can be declared in an Androidmanifest to run in a separate
process. If an application is marked as persistent, it is expected that
all the ProcessRecords created by this application will also be
persistent.

However, processes that run services are currently not marked as
persistent. This patch fixes this issue.

Bug: 355249252
Test: adb shell dumpsys activity processes
Change-Id: I447e7337bead8ed81cd67492484189f0a97faa2b
Signed-off-by: Oleg Lyovin <ovlevin@salutedevices.com>