Revert "Add more conscrypt benchmarks and extend benchmark param..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0c835364-fee0-4e40-8b7a-43798b3363d0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I58921dfc6849bc8232c87f3e3913beed41aff69f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com