Revert "Support getting RPC service in java ServiceManager"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3346909e-2d5c-4d25-8955-c871ebd41d6d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6730cccde0bb612b73886aef41c1c39da7a8e626
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com