Revert "Support getting RPC service in java ServiceManager"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f644da5e-a601-44b8-aa92-873cda2e15ff). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iad484b779a08fe3eaf4f69862635bb0752618b62
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com