Revert "Revert "Revert "Revert "Revert "Workaround build breakag..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/91815f2d-427e-4a82-9317-b685988aeccf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 356856433
Change-Id: Ia55b74669adf976b206bfb05d1868cb22d618ce3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com