Revert "Reapply "New test_module_config_entries for FrameworksCore""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ab4e25b8-5929-4f7d-9ddc-913530e80cb6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I281b3cb1b3aca21fd095a24a1f7654446e5fe7e5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com