Revert "nfc: Eliminate boilder plate code for recovering service..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1bdf2467-c0e3-472b-a704-bc4c98cb6dbb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie84d1231ea76c806018fabf9d22ee4f015bc8349
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com