Revert "Camera: Fix camera id equal check"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1bdf2467-c0e3-472b-a704-bc4c98cb6dbb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I22fff42283f6125850888fc3e89132ebcfc4f53a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com