Don't throw exception when changing unrequested permission state (legacy).
And just log it instead.

This is a simple port of ag/25972914 to the legacypermission subsystem.

Because such an API call is always racy with package changes, especially
when someone is doing things in response to changes (e.g. the role
subsystem). We have to either catch all security exceptions at the call
site, or throw less exceptions while keeping the more important ones. So
not throwing in this case is arguably the better option.

Bug: 321559226
Change-Id: Ia97c7a3d39217b69672a18af5d4c6b76322bfccf
Test: presubmit