Revert "Support getting RPC service in java ServiceManager"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1d4c2f80-c2c5-4f4c-8835-30058b7f0d20). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3bbe1514f6cb5d4d5051bdc95cad68a262d8e761
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com