DO NOT SUBMIT Verify windowing mode and bounds when apply overrides
When apply compat overrides, we need to check the windowing mode before
calculating. If the window is with a fullscreen windowing mode, and does
not overlap with any non-decor insets, we should not exclude the
non-decor insets from its bounds even if the app needs an override.

Bug: 358509380
Bug: 356050755
Test: WindowPolicyTests#testOptOutEdgeToEdgeAppBounds
Test: WindowPolicyTests#testOptOutEdgeToEdgeDisplayMetrics
Flag: EXEMPT bugfix
Change-Id: I2fcdb37a40db088b030ec5aa645da442f5f01f54