Revert "Remove code related to stolen trace sessions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3893b4c0-2144-42eb-b96c-61b4cb605628). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I23d2ab920e8b2198f37ab0a38571530101cb2c87
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com