Revert "nfc: Add New API to get the list of active NFCEE"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ce889392-ab2e-4c45-8610-91268d570994). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia36b13f293572f2e06b58eae802d30e41015317f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com