Revert "nfc: Add New API to get the list of active NFCEE"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/152e4d49-b749-4ffa-b578-be7f073f93d0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0ca29a8574aa26b47c36528ab365f9bc1f8ee0b7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com