Revert "Remove code related to stolen trace sessions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4b81e26-1b70-4aca-9f4c-0a4bc2c2b934). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7a9b3b6de811e0a430607667c10a590e9aa470e2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com